requirements.txt
boto3==1.5.22
botocore==1.8.36
certifi==2018.4.16
chardet==3.0.4
Django==2.0.2
django-filter==1.1.0
djangorestframework==3.8.2
docutils==0.14
idna==2.6
jmespath==0.9.3
Markdown==2.6.11
python-dateutil==2.7.3
pytz==2018.4
requests==2.18.4
s3transfer==0.1.13
simplejson==3.15.0
six==1.11.0
urllib3==1.22
wincertstore==0.2